added security to add edit and deletion of items and login profile views

+++ jewellery/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, get_object_or_404, redirect, reverse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from .models import Item, Item_Cat
 from django.db.models.functions import Lower
@@ -68,8 +69,13 @@ def items_description(request, item_id):
 
     return render(request, 'items/item_description.html', context)
 
+@login_required
 def add_item(request):
     """ Add a product to the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only J Store employees can do that.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ItemForm(request.POST, request.FILES)
         if form.is_valid():
@@ -88,8 +94,13 @@ def add_item(request):
 
     return render(request, template, context)
 
+@login_required
 def edit_item(request, item_id):
     """ Edit a item in the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only J Store employees can do that.')
+        return redirect(reverse('home'))
+    
     item = get_object_or_404(Item, pk=item_id)
     if request.method == 'POST':
         form = ItemForm(request.POST, request.FILES, instance=item)
@@ -111,8 +122,13 @@ def edit_item(request, item_id):
 
     return render(request, template, context)
 
+@login_required
 def delete_item(request, item_id):
     """ Delete a product from the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only J Store employees can do that.')
+        return redirect(reverse('home'))
+    
     item = get_object_or_404(Item, pk=item_id)
     item.delete()
     messages.success(request, 'Item deleted!')
